requirements.txt
disnake
aiosqlite
disnake-paginator
fuzzywuzzy
python-Levenshtein
python-dotenv